requirements.txt
python-dotenv
google-generativeai
google-genai
